requirements/base.txt
tensorflow
asyncio-redis
gensim==4.3.3
mtranslate==2.4.0
transformers==4.36.2
sentencepiece==0.2.0
sacremoses==0.1.1
torch==2.1.2
huggingface_hub==0.20.3
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.0
spacy==3.7.4
alembic==1.11.*
bcrypt==4.0.*
psycopg2-binary==2.9.*
pytest-dotenv==0.5.2
python-jose==3.3.*
SQLAlchemy==2.0.20
SQLAlchemy-Utils==0.37.2
httpx==0.24.*
pytz==2023.3
celery==5.3.*
flower==2.0.*
Jinja2==3.1.*
urllib3==1.26.*
typing-extensions==4.11.0

pandas==2.2.1
openpyxl==3.1.2

pydantic[email]==2.4.*
pydantic-settings==2.0.*
fastapi==0.103.*
uvicorn[standard]==0.23.*
asyncpg==0.29.*
